+++ src/components/Demo.svelte
@@ -7,10 +7,14 @@
   import * as LZString from "lz-string";
   import Editor from "./Editor.svelte";
   import CodeSegmentation from "./CodeSegmentation.svelte";
+  import ColorScheme from "./ColorSchemeEditor.svelte";
+  import { getSystemColorList, toggleTheme, isDark } from "./lightdark.ts";
+
   export let codeGo = "func Example() {\n\tif x {\n\t\treturn\n\t}\n}";
   export let codeC = "void main() {\n\tif (x) {\n\t\treturn;\n\t}\n}";
   export let codePython = "def example():\n    if x:\n        return";
   let offsetToHighlight: number | undefined = undefined;
+  let colorList = getSystemColorList();
   let languages: {
     language: Language;
     text: string;
@@ -36,6 +40,7 @@
   let simplify = true;
   let flatSwitch = false;
   let highlight = true;
+  let colorPicker = false;
   let verbose = urlParams.has("verbose");
   let showSegmentation = urlParams.has("segmentation");
   let debugMode = urlParams.has("debug");
@@ -118,6 +123,23 @@
     const offset = e.detail.offset;
     if (offset !== null && offset !== undefined) editor?.setCursor(offset);
   }
+
+  $: if (!colorPicker && graph) {
+    graph.applyColors(colorList);
+  }
+
+  function onColorPreview(e) {
+    colorList = e.detail.colors;
+    graph.previewColors(colorList);
+  }
+
+  function onToggleClick(e) {
+    toggleTheme();
+  }
+
+  isDark.subscribe(() => {
+    colorList = getSystemColorList();
+  });
 </script>
 
 <main>
@@ -130,46 +152,58 @@
         >VSCode Extension</a
       >
     </div>
+    <div class="themeToggleWrapper">
+      <div class="themeToggle" on:click={onToggleClick}></div>
+    </div>
   </header>
   <div class="editor">
-    <div class="controls">
-      <select bind:value={selection} on:change={(e) => console.log(selection)}>
-        {#each languages as language}
-          <option value={language}>
-            {language.text}
-          </option>
-        {/each}
-      </select>
-      <button on:click={share}>Share (experimental)</button>
-    </div>
-    <div class="codemirror">
-      {#if selection.language === "Go"}
-        <Editor
-          bind:this={editor}
-          bind:code={codeGo}
-          lang={go()}
-          on:cursorMoved={cursorMoved}
-        />
-      {:else if selection.language === "C"}
-        <Editor
-          bind:this={editor}
-          bind:code={codeC}
-          lang={cpp()}
-          on:cursorMoved={cursorMoved}
-        />
-      {:else if selection.language === "Python"}
-        <Editor
-          bind:this={editor}
-          bind:code={codePython}
-          lang={python()}
-          on:cursorMoved={cursorMoved}
-        />
-      {/if}
-    </div>
-    {#if showSegmentation}
-      <div class="segmentation">
-        <CodeSegmentation {code} language={selection.language} {simplify} />
+    {#if colorPicker}
+      <div class="picker">
+        <ColorScheme on:preview={onColorPreview} {colorList} />
       </div>
+    {:else}
+      <div class="controls">
+        <select
+          bind:value={selection}
+          on:change={(e) => console.log(selection)}
+        >
+          {#each languages as language}
+            <option value={language}>
+              {language.text}
+            </option>
+          {/each}
+        </select>
+        <button on:click={share}>Share (experimental)</button>
+      </div>
+      <div class="codemirror">
+        {#if selection.language === "Go"}
+          <Editor
+            bind:this={editor}
+            bind:code={codeGo}
+            lang={go()}
+            on:cursorMoved={cursorMoved}
+          />
+        {:else if selection.language === "C"}
+          <Editor
+            bind:this={editor}
+            bind:code={codeC}
+            lang={cpp()}
+            on:cursorMoved={cursorMoved}
+          />
+        {:else if selection.language === "Python"}
+          <Editor
+            bind:this={editor}
+            bind:code={codePython}
+            lang={python()}
+            on:cursorMoved={cursorMoved}
+          />
+        {/if}
+      </div>
+      {#if showSegmentation}
+        <div class="segmentation">
+          <CodeSegmentation {code} language={selection.language} {simplify} />
+        </div>
+      {/if}
     {/if}
   </div>
 
@@ -181,6 +215,9 @@
 
         <input type="checkbox" id="flatSwitch" bind:checked={flatSwitch} />
         <label for="flatSwitch">Flat Switch</label>
+
+        <input type="checkbox" id="colorPicker" bind:checked={colorPicker} />
+        <label for="colorPicker">Color Picker</label>
 
         {#if debugMode}
           <input type="checkbox" id="verbose" bind:checked={verbose} />
@@ -233,8 +270,8 @@
     grid-area: header;
     font-family: Arial, Helvetica, sans-serif;
     text-align: center;
-    filter: drop-shadow(0 0 0.3rem gray);
-    background-color: white;
+    filter: drop-shadow(0 0 0.3rem var(--panel-shadow-color));
+    background-color: var(--panel-background-color);
     position: relative;
   }
   .controls {
@@ -258,17 +295,17 @@
   .graph,
   .segmentation,
   .editor {
-    background-color: white;
-    filter: drop-shadow(0 0 0.3rem gray);
+    background-color: var(--panel-background-color);
+    filter: drop-shadow(0 0 0.3rem var(--panel-shadow-color));
   }
 
   .links a {
-    color: black;
+    color: var(--text-color);
     padding: 0.5rem;
   }
 
   .links a:hover {
-    background-color: lightskyblue;
+    background-color: var(--link-hover-background);
     padding: 0.5rem;
   }
   .links {
@@ -280,5 +317,28 @@
     right: 3rem;
     top: 0;
     height: 100%;
+  }
+  .themeToggleWrapper {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    position: absolute;
+    top: 0;
+    left: 3rem;
+    height: 100%;
+    line-height: 3rem;
+  }
+  .themeToggle::before {
+    font-size: 2rem;
+    content: var(--theme-toggle-emoji);
+  }
+  .picker {
+    /* position: fixed;
+    z-index: 100;
+    top: 0;
+    left: 0;
+    background-color: white; */
+    font-family: Arial, Helvetica, sans-serif;
   }
 </style>