+++ src/components/Editor.svelte
@@ -3,6 +3,8 @@
   import type { EditorView } from "@codemirror/view";
   import { createEventDispatcher } from "svelte";
   import CodeMirror from "svelte-codemirror-editor";
+  import { oneDark } from "@codemirror/theme-one-dark";
+  import { isDark } from "./lightdark.ts";
   export let code: string;
   export let lang: LanguageSupport;
 
@@ -55,4 +57,5 @@
   tabSize={4}
   lineWrapping={true}
   on:ready={onEditorReady}
+  theme={$isDark ? oneDark : undefined}
 />